requirements.txt
# Core dependencies
torch>=1.12.0
transformers>=4.20.0
numpy>=1.21.0
pandas>=1.4.0

# Bioinformatics
biopython>=1.79

# Progress bars and utilities
tqdm>=4.64.0

# Optional: for ESM-fold integration
fair-esm>=2.0.0

# Optional: for structure prediction
colabfold>=1.5.0

# Optional: for homology detection
hhblits>=3.3.0

# Optional: for structure analysis
foldseek>=7.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.7.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Additional utilities
pyyaml>=6.0
click>=8.0.0
joblib>=1.1.0
scikit-learn>=1.1.0

# For data handling
h5py>=3.7.0
tables>=3.7.0

# Optional: for protein structure analysis
biotite>=0.36.0
mdanalysis>=2.0.0

# Optional: for deep learning optimizations
apex>=0.1
flash-attn>=0.2.0

# Optional: for distributed training
accelerate>=0.15.0